Many Various Changes (#3)
* Fix fallback not applying modifiers correctly
* Fix bad default font size and improve tab actions
* Allow changing increment in numeric up down
* Fix menu using wrong position to open to in Open() function
* Add all other sizers to WindowCanvas
* Fix dx and dy for Windows in mouse move
* Better handle dragger selection
* Reflow text on a scale change
* Update glew
* Improve sleep with animations
* Handle layout with runtime font size changes better for many controls
* Handle font size scaling well in linux
* Handle font scale better
* Fix x11 unicode pasting
* Fix a missing argument that prevented fontstash from working at all with
truetype
* Make sure color updates before change event fires in colorselector
* Fix numeric up down editing near min/max values
* Fix Layout Glitch For Tree Node
* Fix some of window resizing not taking into account scale
* Fix Color Picker Property Button Menu Layout, Fix Combo Box Property Dropdowns Not Taking Into Account Scrolling
* Fix some x11 crashes/glitches
* Move color functions to utility
* Fix float/numeric properties not always returning right value on update
* Fix syntax error due to probably poor commit
* Fix editing float numeric up down
* Add missing wglew header
* Fix tooltips not responding to font scale changes
* Remove pointless line
* Fixup numeric up down more
* Add onFocusGained to buttons and outline to color buttons
* Fix disabling resizing missing a few sizers
* Embed default skin inside Gwen applications
* Many improvements/tweaks
* Add support for "menu" windows that dont show up in taskbar
* Fix some memory leaks and initialization order warnings/bugs
* Add ShouldRedrawOnHover virtual function to Base control
* Greatly reduce number of pointless re-renderers by only redrawing on necessary hovered control changes
* Fix menus not laying out properly with different font scales
* Add function to get selected nodes in a tree
* Allow proper multi-selection in trees
* Update hovered control each render to properly catch layout updates
* Make tooltip disappear after mouse movement and prevent it causing unnecessary redraws
* Add missing default skin resource
* Add function to delete all children to base control
* Don't count right clicks on buttons as up/down
* Allow changing title of WindowCanvases
* Add override to virtual function overrides
* Fix space event not being called on linux
* Allow setting texture of image panel directly and add keep aspect ratio option
* Properly redraw dragger after click/release
* Handle docked items in scrollcontrol properly
* Add context menu to textbox
* Properly hide cursor when selecting in textbox
* Allow enabling vsync
* Change slider to double from float
* Fix keypad enter on linux
* Add collapse all to nodes
* Fix linux build with wgl

+++ gwen/src/Controls/HorizontalSlider.cpp
@@ -18,9 +18,9 @@ GWEN_CONTROL_CONSTRUCTOR( HorizontalSlider )
 	m_SliderBar->SetHorizontal( true );
 }
 
-float HorizontalSlider::CalculateValue()
+double HorizontalSlider::CalculateValue()
 {
-	return ( float ) m_SliderBar->X() / ( float )( Width() - m_SliderBar->Width() );
+	return ( double ) m_SliderBar->X() / ( double )( Width() - m_SliderBar->Width() );
 }
 
 void HorizontalSlider::UpdateBarFromValue()
